+++ controllers/version_retreiver_controller.js
@@ -1,0 +1,9 @@
+module.exports = class {
+  constructor(versionRetreiver) {
+    this.versionRetreiver = versionRetreiver
+  }
+
+  async index(ctx) {
+    ctx.body = await this.versionRetreiver.getVersion()
+  }
+}