+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.6
 - Add CodeClimate code coverage to CircleCI
+- Implement `versionRetreiver` service and `/_version` API
 
 ## 0.7.5
 - Support for `prefix` and `global tags` in StatsD configuration